do relation homework

+++ app/Http/Controllers/Api/CategoryController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Category;
 use App\ResponseHelper;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class CategoryController extends Controller
 {
@@ -14,9 +15,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        // $categories =  Category::all();
-        // $categories =  Category::withAvg('books' , 'price')->get();
-        $categories =  Category::withCount('books')->get();
+        $categories =  Category::all();
        return ResponseHelper::success(' جميع الأصناف',$categories);
     }
 
@@ -26,10 +25,19 @@ class CategoryController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'name' => 'required|max:50|unique:categories'
+            'name' => 'required|max:50|unique:categories',
+            'image' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048'
         ]);
         $category = new Category();
         $category->name = $request->name;
+
+        if ($request->hasFile('image')) {
+            $file = $request->file('image');
+            $filename = "$request->id." . $file->extension();
+            Storage::putFileAs('category-images', $file, $filename);
+            $category->image = $filename;
+        }
+
         $category->save();
         return ResponseHelper::success("تمت إضافة الصنف" , $category);
     }
@@ -58,6 +66,13 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::find($id);
+        
+        // Check if category has associated books
+        if ($category->books()->count() > 0) {
+            return ResponseHelper::error("لا يمكن حذف الصنف لوجود كتب مرتبطة به" . $category->books()->count() , null, 409);
+        
+        }
+        
         $category->delete();
         return ResponseHelper::success("تم حذف الصنف" , $category);
     }
